requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
requests

openai
sentence-transformers
scikit-learn
numpy

mediapipe
opencv-python
opencv-contrib-python
sounddevice
soundfile

pdfplumber
PyPDF2
pypdfium2

torch
transformers
sentencepiece
safetensors
tqdm

websockets
aiohttp
librosa
python-multipart
